refactor/#86: profile module facade pattern 적용
