+++ src/main/java/weteam/backend/domain/profile/ProfileService.java
@@ -1,34 +1,22 @@
 package weteam.backend.domain.profile;
 
 import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import weteam.backend.application.auth.SecurityUtil;
-import weteam.backend.application.handler.exception.CustomException;
-import weteam.backend.application.handler.exception.ErrorCode;
+import weteam.backend.domain.profile.entity.ProfileImage;
 import weteam.backend.domain.user.UserRepository;
 import weteam.backend.domain.user.entity.User;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class ProfileService {
   private final ProfileRepository profileRepository;
-  private final UserRepository userRepository;
-  private final SecurityUtil securityUtil;
 
   @Transactional
-  public void addProfile(final Long imageIdx) {
-    final User user = userRepository.findById(securityUtil.getCurrentUserId()).orElseThrow(CustomException.raise(ErrorCode.NOT_FOUND));
-    if (user.getProfileImage() != null) {
-      throw new CustomException(ErrorCode.DUPLICATE);
-    }
+  public void addProfile(final Long imageIdx, final User user) {
     final ProfileImage image = ProfileImage.from(imageIdx, user);
     profileRepository.save(image);
-  }
-
-  @Transactional
-  public void updateProfile(final Long imageIdx) {
-    ProfileImage image = profileRepository.findByUserId(securityUtil.getCurrentUserId()).orElseThrow(CustomException.raise(ErrorCode.NOT_FOUND));
-    image.updateImage(imageIdx);
   }
 }
