+++ src/main/java/weteam/backend/domain/user/entity/User.java
@@ -9,7 +9,7 @@ import weteam.backend.application.BaseEntity;
 import weteam.backend.domain.alarm.entity.Alarm;
 import weteam.backend.domain.meeting.entity.Meeting;
 import weteam.backend.domain.meeting.entity.MeetingUser;
-import weteam.backend.domain.profile.ProfileImage;
+import weteam.backend.domain.profile.entity.ProfileImage;
 import weteam.backend.domain.project.entity.Project;
 import weteam.backend.domain.project.entity.ProjectUser;
 import weteam.backend.domain.user.dto.RequestUserDto;
@@ -75,6 +75,10 @@ public class User extends BaseEntity {
 
   public void updateDevice(final String deviceToken) {
     this.deviceToken = deviceToken;
+  }
+
+  public void updateProfileImage(final long imageIdx) {
+    this.profileImage.updateImage(imageIdx);
   }
 
   public void toggleReceivePermission() {
